Se cierra backend y detalles del site.

+++ Backend/Controllers/TiendaVirtualController.cs
@@ -44,6 +44,7 @@ namespace Backend.Controllers
         {
             var idUser = User.Identity.GetUserId();
             var tiendas = _bl.ObtenerTiendaDelUsuario(idUser);
+            ViewBag.numero = tiendas.Count();
             return View(tiendas);
         }
 
@@ -75,7 +76,10 @@ namespace Backend.Controllers
             {
                  var idUser = User.Identity.GetUserId();
                  var tienda= _bl.ObtenerTiendaDelUsuario(User.Identity.GetUserName());
-
+                 var tiendanumber = _bl.ObtenerTiendaDelUsuario(idUser).Count();
+                 
+                 return View();
+                 //return Content("<script language='javascript' type='text/javascript'>mensaje('Tienda Virtual', 'Solo se permite crear una tienda virtual por usuario', 'error');</script>"); }
                     
             }
             catch (Exception)
@@ -85,7 +89,7 @@ namespace Backend.Controllers
             }
             
             
-            return View();
+            
         }
 
         // POST: TiendaVirtual/Create
@@ -320,7 +324,7 @@ namespace Backend.Controllers
 
 
         [Authorize(Roles = "Usuario")]
-        public ActionResult Upload(int id)
+        public ActionResult Upload(int id, string mensaje,string tipo)
         {
             TiendaVirtualDTO tienda = _bl.ObtenerTienda(id);
             //CloudBlobContainer blobContainer = _bss.GetContainerTienda(tienda.Dominio);
@@ -331,7 +335,11 @@ namespace Backend.Controllers
             //{
             //    blobs.Add(blob.Uri.ToString());
             //}
-
+            if (mensaje == null) mensaje = "";
+            if (tipo == null) tipo = "";
+            ViewBag.Imagenes = _bl.ObtenerImagenes(id);
+            ViewBag.mensaje = mensaje;
+            ViewBag.tipo = tipo;
             return View(tienda);
         }
 
@@ -339,28 +347,37 @@ namespace Backend.Controllers
         [Authorize(Roles = "Usuario")]
         public ActionResult Upload(TiendaVirtualDTO tienda, HttpPostedFileBase image)
         {
-
-                TiendaVirtualDTO tiendaDTO = _bl.ObtenerTienda(tienda.TiendaVitualId);
+            TiendaVirtualDTO tiendaDTO = _bl.ObtenerTienda(tienda.TiendaVitualId);
+            if (image.ContentLength < 2097152)
+            {
+                
                 var nombreFoto = tienda.Dominio + Guid.NewGuid().ToString() + "_portada";
                 CloudBlobContainer blobContainer = _bss.GetContainerTienda(tienda.Dominio);
                 CloudBlockBlob blob = blobContainer.GetBlockBlobReference(nombreFoto);
-                
+
                 ImagenesDTO imagen = new ImagenesDTO();
                 imagen.TiendaId = tiendaDTO.TiendaVitualId;
                 imagen.Nombre = nombreFoto;
                 imagen.UrlImagenMediana = blob.Uri.ToString();
                 _bl.AgregarImagenTienda(imagen);
-                
+
 
                 blob.UploadFromStream(image.InputStream);
 
-                return RedirectToAction("Upload", "TiendaVirtual", new { id = tiendaDTO.TiendaVitualId });
+                return RedirectToAction("Upload", "TiendaVirtual", new { id = tiendaDTO.TiendaVitualId, mensaje = "Se agrego la imagen correctamente", tipo = "success" });
+            }
+            else
+            {
+                return RedirectToAction("Upload", "TiendaVirtual", new { id = tiendaDTO.TiendaVitualId, mensaje = "Tamaño maximo 2 MB", tipo="warning" });
+               
+            }
+  
         }
 
 
         [HttpPost]
         [Authorize(Roles = "Usuario")]
-        public string DeleteImage(string Name, string Uri, int id)
+        public ActionResult DeleteImage(string Name, string Uri, int id)
         {
             TiendaVirtualDTO tienda = _bl.ObtenerTienda(id);
 
@@ -369,11 +386,24 @@ namespace Backend.Controllers
             CloudBlobContainer blobContainer = _bss.GetContainerTienda(tienda.Dominio);
             CloudBlockBlob blob = blobContainer.GetBlockBlobReference(filename);
 
-            _bl.EliminarImagenTienda(tienda.TiendaVitualId, Name);
+            bool ok=_bl.EliminarImagenTienda(tienda.TiendaVitualId, Name);
 
-            blob.Delete();
+            if (ok)
+            {
+                blob.Delete();
+                ViewBag.mensaje = "Prueba";
+                ViewBag.tipo = "Prueba";
+                return Content("<script language='javascript' type='text/javascript'>mensaje('Tienda Virtual', 'Solo se permite crear una tienda virtual por usuario', 'error');</script>"); 
+                    
+            }
+            else 
+            {
+                ViewBag.mensaje = "";
+                ViewBag.tipo = "";
+                return Content("<script language='javascript' type='text/javascript'>mensaje('Tienda Virtual', 'Solo se permite crear una tienda virtual por usuario', 'error');</script>");
+            }
 
-            return "File delete";
+            
         }
 
         [Authorize(Roles = "Admin")]
@@ -417,6 +447,7 @@ namespace Backend.Controllers
         {
             List<SubastaAux> subastas;
             List<Usuario> usuarios;
+
 
             IEnumerable<UsuarioReporte> modelList = new List<UsuarioReporte>();
             IEnumerable<SubastaReporte> modelList2 = new List<SubastaReporte>();
@@ -516,8 +547,7 @@ namespace Backend.Controllers
             IEnumerable<ReporteLineal> modelList2 = new List<ReporteLineal>();
 
             TipoReporte e = (TipoReporte)Enum.Parse(typeof(TipoReporte), "Usuario");
-            //DateTime fechai = Convert.ToDateTime(fechaini);
-            //DateTime fechaf = Convert.ToDateTime(fechafin);
+
 
             if (rep.tipo.Equals(e))
             {
@@ -620,6 +650,92 @@ namespace Backend.Controllers
 
 
             //return Json(modelList, JsonRequestBehavior.AllowGet);
+
+        }
+
+        [Authorize(Roles = "Admin")]
+        public ActionResult ReportesSubastas()
+        {
+
+            // ViewData["Tiendas"] = _bl.ObtenerTiendas().ToList();
+
+
+
+            ViewBag.tiendas = _bl.ObtenerTiendas().ToList();
+            return View();
+        }
+
+        [Authorize(Roles = "Admin")]
+        public ActionResult ReportesUsuarios()
+        {
+
+            // ViewData["Tiendas"] = _bl.ObtenerTiendas().ToList();
+
+
+
+            ViewBag.tiendas = _bl.ObtenerTiendas().ToList();
+            return View();
+        }
+
+        public ActionResult ReportesDetalle(Reporte rep)
+        {
+            List<SubastaAux> subastas;
+            List<UsuarioAux> usuarios;
+
+
+            IEnumerable<UsuarioReporte> modelList = new List<UsuarioReporte>();
+            IEnumerable<SubastaReporte> modelList2 = new List<SubastaReporte>();
+
+            TipoReporte e = (TipoReporte)Enum.Parse(typeof(TipoReporte), "Usuario");
+            //DateTime fechai = Convert.ToDateTime(fechaini);
+            //DateTime fechaf = Convert.ToDateTime(fechafin);
+
+            if (rep.tipo.Equals(e))
+            {
+                usuarios = _bl.DetUsers(rep.dominio, rep.fechaini).ToList();
+                modelList = usuarios.Select(x =>
+                                            new UsuarioReporte()
+                                            {
+                                                tipo = "Usuario",
+                                                nick = x.nick,
+                                                nombre = x.nombre,
+                                                apellido = x.apellido,
+                                                email = x.email,
+                                                fecha_Alta = x.fecha_Alta
+
+                                            });
+                return Json(modelList, JsonRequestBehavior.AllowGet);
+
+
+            }
+            else
+            {
+                subastas = _bl.DetSub(rep.dominio, rep.fechaini).ToList();
+                modelList2 = subastas.Select(x =>
+                                            new SubastaReporte()
+                                            {
+                                                tipo = "Subasta",
+                                                titulo = x.titulo,
+                                                nickComprador = x.nombreComprador,
+                                                nickVendedor = x.nombreVendedor,
+                                                precio_Base = x.precio_Base,
+                                                fecha_Inicio = x.fecha_Inicio
+
+                                            });
+                return Json(modelList2, JsonRequestBehavior.AllowGet);
+            }
+
+        }
+
+
+        public ActionResult ExisteDominio(string dominio)
+        {
+            string retorno;
+            bool existe = _bl.ExisteDominio(dominio);
+            if(existe) retorno="true";
+            else retorno="false";
+
+            return Json(retorno, JsonRequestBehavior.AllowGet);
 
         }
     }
